+++ docs/source_code_templates/c_cpp_template.c
@@ -1,0 +1,8 @@
+/** @file module.c
+ *
+ * @brief A one line summary of the module or program
+ *
+ * Created by IRNAS, more info at: https://www.irnas.eu/
+ */
+
+#include "module.h"