+++ docs/source_code_templates/python_template.py
@@ -1,0 +1,13 @@
+#!/usr/bin/env python3
+"""A one line summary of the module or program
+
+Leave one blank line.  The rest of this docstring should contain an
+overall description of the module or program.  Optionally, it may also
+contain a brief description of exported classes and functions and/or usage
+examples.
+
+  Typical usage example:
+
+  foo = ClassFoo()
+  bar = foo.FunctionBar()
+"""